requirements.txt
scipy==0.19.1
numpy==1.13.3
apscheduler==3.4.0
pandas==0.20.3
pymongo==3.5.1
raven==6.5.0
hvac[parser]==0.2.16
pytimeparse==1.1.5
requests==2.8.1
testfixtures==5.4.0
prometheus_client
grpcio-tools==1.12.1
statsmodels==0.9.0
pytest-catchlog==1.2.2

